backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
neo4j==5.26.0
pydantic-settings==2.5.2
python-dotenv==1.0.1
